refactor(jellyfin): move track preference and runtime switching to the service
Language preference resolution, queue stream retargeting, mpv track set
attempts, runtime selected-stream pickers, and now-playing metadata
merge/enrich now live in jellyfin_service. The queue retarget write moves
with it, so the transition-inventory QUEUE pin gains the service as a
documented content-management exception.
