+++ app/relaytv_app/routes/__init__.py
@@ -2346,6 +2346,18 @@ _jellyfin_auto_prefers_transcode = jellyfin_service.auto_prefers_transcode
 _select_jellyfin_playback_url = jellyfin_service.select_playback_url
 _first_playable_jellyfin_episode = jellyfin_service.first_playable_episode
 _resolve_jellyfin_playable_item = jellyfin_service.resolve_playable_item
+_normalize_lang_pref = jellyfin_service._normalize_lang_pref
+_language_aliases = jellyfin_service._language_aliases
+_language_matches = jellyfin_service._language_matches
+_preferred_jellyfin_stream_indices = jellyfin_service.preferred_stream_indices
+_retarget_jellyfin_queue_stream_preferences = jellyfin_service.retarget_queue_stream_preferences
+_is_generic_playback_title = jellyfin_service._is_generic_playback_title
+_merge_jellyfin_playback_metadata = jellyfin_service.merge_playback_metadata
+_jellyfin_enrich_now_stream_metadata = jellyfin_service.enrich_now_stream_metadata
+_jellyfin_try_set_mpv_audio_track = jellyfin_service.try_set_mpv_audio_track
+_jellyfin_try_set_mpv_subtitle_track = jellyfin_service.try_set_mpv_subtitle_track
+_jellyfin_runtime_selected_audio_stream = jellyfin_service.runtime_selected_audio_stream
+_jellyfin_runtime_selected_subtitle_stream = jellyfin_service.runtime_selected_subtitle_stream
 
 
 def _reset_jellyfin_command_state() -> None:
@@ -2432,182 +2444,6 @@ def _env_flag(name: str, *, default: bool = False) -> bool:
     if raw is None or str(raw).strip() == "":
         return bool(default)
     return str(raw).strip().lower() in ("1", "true", "yes", "on")
-
-
-def _normalize_lang_pref(raw: str) -> str:
-    text = str(raw or "").strip().lower().replace("_", "-")
-    if "," in text:
-        text = text.split(",", 1)[0].strip()
-    return text
-
-
-def _language_aliases(raw: str) -> set[str]:
-    base = _normalize_lang_pref(raw)
-    if not base:
-        return set()
-    aliases: set[str] = {base}
-    # Normalize common locale stems.
-    if "-" in base:
-        aliases.add(base.split("-", 1)[0])
-    # Map common ISO-639-2 and alternate spellings to ISO-639-1 stems.
-    to_stem = {
-        "eng": "en",
-        "por": "pt",
-        "spa": "es",
-        "jpn": "ja",
-        "deu": "de",
-        "ger": "de",
-        "fra": "fr",
-        "fre": "fr",
-        "ita": "it",
-        "rus": "ru",
-        "kor": "ko",
-        "zho": "zh",
-        "chi": "zh",
-        "ara": "ar",
-        "ces": "cs",
-        "cze": "cs",
-        "nld": "nl",
-        "dut": "nl",
-        "pol": "pl",
-        "tur": "tr",
-        "hun": "hu",
-        "dan": "da",
-        "fin": "fi",
-        "ron": "ro",
-        "rum": "ro",
-        "swe": "sv",
-        "ell": "el",
-        "gre": "el",
-        "nob": "nb",
-    }
-    stem = to_stem.get(base, "")
-    if stem:
-        aliases.add(stem)
-    short = base.split("-", 1)[0]
-    if short in to_stem:
-        aliases.add(to_stem[short])
-    # Regional variants used by stream metadata.
-    if "pt" in aliases:
-        aliases.update({"pt-br", "pt-pt"})
-    if "en" in aliases:
-        aliases.update({"en-us", "en-gb"})
-    if "es" in aliases:
-        aliases.update({"es-419", "es-es"})
-    return {a for a in aliases if a}
-
-
-def _language_matches(pref: str, candidate: str) -> bool:
-    p_alias = _language_aliases(pref)
-    c_alias = _language_aliases(candidate)
-    if not p_alias or not c_alias:
-        return False
-    if p_alias.intersection(c_alias):
-        return True
-    return False
-
-
-def _preferred_jellyfin_stream_indices(item_id: str) -> tuple[str, str]:
-    iid = str(item_id or "").strip()
-    if not iid:
-        return "", ""
-    settings = state.get_settings() if hasattr(state, "get_settings") else {}
-    audio_pref = _normalize_lang_pref(str(settings.get("jellyfin_audio_lang") or ""))
-    sub_pref = _normalize_lang_pref(str(settings.get("jellyfin_sub_lang") or ""))
-    if not audio_pref and not sub_pref:
-        return "", ""
-    sub_off = sub_pref in {"off", "none", "disabled", "no", "false", "0"}
-    try:
-        detail = jellyfin_receiver.get_item_detail(iid)
-    except Exception:
-        detail = {}
-    audio_streams = detail.get("audio_streams") if isinstance(detail, dict) else []
-    subtitle_streams = detail.get("subtitle_streams") if isinstance(detail, dict) else []
-    audio_idx = ""
-    sub_idx = "-1" if sub_off else ""
-    if audio_pref and isinstance(audio_streams, list):
-        for stream in audio_streams:
-            if not isinstance(stream, dict):
-                continue
-            if _language_matches(audio_pref, str(stream.get("language") or "")):
-                try:
-                    audio_idx = str(int(stream.get("index")))
-                except Exception:
-                    audio_idx = ""
-                if audio_idx:
-                    break
-    if sub_pref and (not sub_off) and isinstance(subtitle_streams, list):
-        for stream in subtitle_streams:
-            if not isinstance(stream, dict):
-                continue
-            if _language_matches(sub_pref, str(stream.get("language") or "")):
-                try:
-                    sub_idx = str(int(stream.get("index")))
-                except Exception:
-                    sub_idx = ""
-                if sub_idx:
-                    break
-    return audio_idx, sub_idx
-
-
-def _retarget_jellyfin_queue_stream_preferences() -> int:
-    """Best-effort: rewrite queued Jellyfin URLs using current language prefs."""
-    with state.QUEUE_LOCK:
-        snapshot = list(state.QUEUE)
-    if not snapshot:
-        return 0
-
-    changed = 0
-    updated_queue: list[object] = list(snapshot)
-    for i, entry in enumerate(snapshot):
-        if not isinstance(entry, dict):
-            continue
-        raw_url = str(entry.get("url") or "").strip()
-        if not raw_url:
-            continue
-        provider = str(entry.get("provider") or "").strip().lower()
-        item_id = str(entry.get("jellyfin_item_id") or "").strip()
-        if not item_id:
-            item_id = _extract_jellyfin_item_id_from_url_raw(raw_url)
-        if provider != "jellyfin" and not item_id:
-            continue
-        pref_audio_idx, pref_sub_idx = _preferred_jellyfin_stream_indices(item_id) if item_id else ("", "")
-        if pref_audio_idx == "" and pref_sub_idx == "":
-            continue
-        next_url = _apply_jellyfin_stream_params(
-            raw_url,
-            audio_stream_index=pref_audio_idx,
-            subtitle_stream_index=pref_sub_idx,
-        )
-        media_source_id = _first_nonempty_str(
-            [
-                str(entry.get("jellyfin_media_source_id") or "").strip(),
-                _extract_jellyfin_media_source_id_from_url(raw_url),
-            ]
-        )
-        next_url = _apply_jellyfin_media_source_param(next_url, media_source_id=media_source_id)
-        if next_url and next_url != raw_url:
-            out_entry = dict(entry)
-            out_entry["url"] = next_url
-            if item_id:
-                out_entry["jellyfin_item_id"] = item_id
-            updated_queue[i] = out_entry
-            changed += 1
-
-    if changed <= 0:
-        return 0
-
-    with state.QUEUE_LOCK:
-        state.QUEUE[:] = updated_queue
-    try:
-        state.persist_queue()
-    except Exception:
-        pass
-    try:
-        player.prime_mpv_up_next_from_queue(force=True)
-    except Exception:
-        pass
-    return changed
 
 
 def _extract_api_key_from_url(url: str) -> str:
@@ -2713,516 +2549,6 @@ def _smart_item_from_url(url: str, *, start_pos: float | None = None, lightweigh
     except TypeError:
         # Compatibility for tests/patches that mock make_item(url) without kwargs.
         return resolver.make_item(shared)
-
-
-def _is_generic_playback_title(title: object, url: object) -> bool:
-    t = str(title or "").strip()
-    if not t:
-        return True
-    low = t.lower()
-    if low in {"stream", "download", "video", "playback", "master", "master.m3u8", "main", "main.m3u8"}:
-        return True
-    u = str(url or "").strip()
-    if u and t == u:
-        return True
-    return False
-
-
-def _merge_jellyfin_playback_metadata(now: dict, enriched: dict) -> dict:
-    out = dict(now)
-    provider = str(enriched.get("provider") or "").strip().lower()
-    if provider == "jellyfin" and str(out.get("provider") or "").strip().lower() != "jellyfin":
-        out["provider"] = "jellyfin"
-
-    if enriched.get("title") and _is_generic_playback_title(out.get("title"), out.get("url")):
-        out["title"] = enriched.get("title")
-
-    if enriched.get("channel") and not out.get("channel"):
-        out["channel"] = enriched.get("channel")
-    if enriched.get("thumbnail") and not out.get("thumbnail"):
-        out["thumbnail"] = enriched.get("thumbnail")
-    if enriched.get("thumbnail_local") and not out.get("thumbnail_local"):
-        out["thumbnail_local"] = enriched.get("thumbnail_local")
-
-    # Keep canonical Jellyfin identifiers from enriched metadata when present.
-    if enriched.get("jellyfin_item_id"):
-        out["jellyfin_item_id"] = enriched.get("jellyfin_item_id")
-    if enriched.get("jellyfin_media_source_id"):
-        out["jellyfin_media_source_id"] = enriched.get("jellyfin_media_source_id")
-    return out
-
-
-def _jellyfin_enrich_now_stream_metadata(
-    now: dict,
-    *,
-    detail: dict | None = None,
-    audio_stream_index: str = "",
-    subtitle_stream_index: str = "",
-) -> dict:
-    out = dict(now or {})
-    meta = detail if isinstance(detail, dict) else {}
-    audio_streams = meta.get("audio_streams") if isinstance(meta.get("audio_streams"), list) else []
-    subtitle_streams = meta.get("subtitle_streams") if isinstance(meta.get("subtitle_streams"), list) else []
-    if audio_streams:
-        out["audio_streams"] = audio_streams
-    if subtitle_streams:
-        out["subtitle_streams"] = subtitle_streams
-
-    selected_audio = str(audio_stream_index or out.get("jellyfin_audio_stream_index") or "").strip()
-    if not selected_audio:
-        selected_audio = _extract_jellyfin_audio_stream_index_from_url(str(out.get("url") or ""))
-    selected_sub = str(subtitle_stream_index or out.get("jellyfin_subtitle_stream_index") or "").strip()
-    if not selected_sub:
-        selected_sub = _extract_jellyfin_subtitle_stream_index_from_url(str(out.get("url") or ""))
-
-    selected_audio_lang = ""
-    if isinstance(audio_streams, list):
-        target_idx = None
-        try:
-            target_idx = int(selected_audio) if selected_audio != "" else None
-        except Exception:
-            target_idx = None
-        if target_idx is not None:
-            for row in audio_streams:
-                if not isinstance(row, dict):
-                    continue
-                try:
-                    if int(row.get("index")) == target_idx:
-                        selected_audio_lang = str(row.get("language") or "").strip()
-                        break
-                except Exception:
-                    continue
-        if not selected_audio_lang:
-            for row in audio_streams:
-                if isinstance(row, dict) and bool(row.get("is_default")) and str(row.get("language") or "").strip():
-                    selected_audio_lang = str(row.get("language") or "").strip()
-                    break
-
-    selected_sub_lang = "off" if selected_sub == "-1" else ""
-    if isinstance(subtitle_streams, list) and selected_sub != "-1":
-        target_sub_idx = None
-        try:
-            target_sub_idx = int(selected_sub) if selected_sub != "" else None
-        except Exception:
-            target_sub_idx = None
-        if target_sub_idx is not None:
-            for row in subtitle_streams:
-                if not isinstance(row, dict):
-                    continue
-                try:
-                    if int(row.get("index")) == target_sub_idx:
-                        selected_sub_lang = str(row.get("language") or "").strip()
-                        break
-                except Exception:
-                    continue
-        if not selected_sub_lang:
-            for row in subtitle_streams:
-                if isinstance(row, dict) and bool(row.get("is_default")) and str(row.get("language") or "").strip():
-                    selected_sub_lang = str(row.get("language") or "").strip()
-                    break
-
-    if selected_audio != "":
-        out["jellyfin_audio_stream_index"] = selected_audio
-    if selected_sub != "":
-        out["jellyfin_subtitle_stream_index"] = selected_sub
-
-    out["audio_language"] = selected_audio_lang or str(meta.get("audio_language") or out.get("audio_language") or "").strip()
-    out["subtitle_language"] = selected_sub_lang or str(meta.get("subtitle_language") or out.get("subtitle_language") or "").strip()
-    out["jellyfin_audio_language"] = str(out.get("audio_language") or "").strip()
-    out["jellyfin_subtitle_language"] = str(out.get("subtitle_language") or "").strip()
-    return out
-
-
-def _jellyfin_try_set_mpv_audio_track(
-    *,
-    language: str = "",
-    display: str = "",
-    preferred_stream_index: int | None = None,
-) -> bool:
-    target_lang = _normalize_lang_pref(str(language or ""))
-    target_display = str(display or "").strip().lower()
-    target_display_tokens = [tok for tok in re.split(r"[^a-z0-9]+", target_display) if len(tok) >= 3]
-    try:
-        track_list = player.mpv_get("track-list")
-    except Exception:
-        track_list = None
-    if not isinstance(track_list, list):
-        return False
-
-    candidates: list[tuple[int, int]] = []
-    for idx, row in enumerate(track_list):
-        if not isinstance(row, dict):
-            continue
-        if str(row.get("type") or "").strip().lower() != "audio":
-            continue
-        try:
-            tid = int(row.get("id"))
-        except Exception:
-            continue
-        if tid <= 0:
-            continue
-        src_id = None
-        ff_index = None
-        try:
-            src_id = int(row.get("src-id"))
-        except Exception:
-            src_id = None
-        try:
-            ff_index = int(row.get("ff-index"))
-        except Exception:
-            ff_index = None
-        lang = _normalize_lang_pref(str(row.get("lang") or row.get("language") or ""))
-        title = str(row.get("title") or row.get("name") or "").strip().lower()
-        score = 0
-        if preferred_stream_index is not None:
-            if ff_index is not None and ff_index == preferred_stream_index:
-                score += 20
-            if src_id is not None and (src_id - 1) == preferred_stream_index:
-                score += 18
-        if target_lang and _language_matches(target_lang, lang):
-            score += 6
-        if target_display and title:
-            if target_display in title or title in target_display:
-                score += 4
-            elif target_display_tokens:
-                token_hits = sum(1 for tok in target_display_tokens if tok in title)
-                score += min(3, token_hits)
-        if score <= 0:
-            continue
-        # Stable tie-break by track order.
-        candidates.append((score * 1000 - idx, tid))
-    if not candidates:
-        return False
-    candidates.sort(reverse=True)
-    selected_tid = int(candidates[0][1])
-
-    try:
-        player.mpv_set("aid", selected_tid)
-    except Exception:
-        return False
-
-    # Confirm selection if possible.
-    try:
-        updated = player.mpv_get("track-list")
-    except Exception:
-        updated = None
-    if isinstance(updated, list):
-        for row in updated:
-            if not isinstance(row, dict):
-                continue
-            if str(row.get("type") or "").strip().lower() != "audio":
-                continue
-            if not bool(row.get("selected")):
-                continue
-            src_id = None
-            ff_index = None
-            try:
-                src_id = int(row.get("src-id"))
-            except Exception:
-                src_id = None
-            try:
-                ff_index = int(row.get("ff-index"))
-            except Exception:
-                ff_index = None
-            if preferred_stream_index is not None:
-                if ff_index is not None and ff_index == preferred_stream_index:
-                    return True
-                if src_id is not None and (src_id - 1) == preferred_stream_index:
-                    return True
-            lang = _normalize_lang_pref(str(row.get("lang") or row.get("language") or ""))
-            title = str(row.get("title") or row.get("name") or "").strip().lower()
-            if target_lang and _language_matches(target_lang, lang):
-                return True
-            if target_display and target_display in title:
-                return True
-    return False
-
-
-def _jellyfin_try_set_mpv_subtitle_track(
-    *,
-    language: str = "",
-    display: str = "",
-    preferred_stream_index: int | None = None,
-    off: bool = False,
-) -> bool:
-    if off:
-        try:
-            player.mpv_set("sid", "no")
-            try:
-                player.mpv_set("sub-visibility", False)
-            except Exception:
-                pass
-            return True
-        except Exception:
-            return False
-
-    target_lang = _normalize_lang_pref(str(language or ""))
-    target_display = str(display or "").strip().lower()
-    target_display_tokens = [tok for tok in re.split(r"[^a-z0-9]+", target_display) if len(tok) >= 3]
-    try:
-        track_list = player.mpv_get("track-list")
-    except Exception:
-        track_list = None
-    if not isinstance(track_list, list):
-        return False
-
-    candidates: list[tuple[int, int]] = []
-    for idx, row in enumerate(track_list):
-        if not isinstance(row, dict):
-            continue
-        if not _jellyfin_track_type_is_subtitle(row.get("type")):
-            continue
-        try:
-            tid = int(row.get("id"))
-        except Exception:
-            continue
-        if tid <= 0:
-            continue
-        src_id = None
-        ff_index = None
-        try:
-            src_id = int(row.get("src-id"))
-        except Exception:
-            src_id = None
-        try:
-            ff_index = int(row.get("ff-index"))
-        except Exception:
-            ff_index = None
-        lang = _normalize_lang_pref(str(row.get("lang") or row.get("language") or ""))
-        title = str(row.get("title") or row.get("name") or "").strip().lower()
-        score = 0
-        if preferred_stream_index is not None:
-            if ff_index is not None and ff_index == preferred_stream_index:
-                score += 20
-            if src_id is not None and (src_id - 1) == preferred_stream_index:
-                score += 18
-        if target_lang and _language_matches(target_lang, lang):
-            score += 6
-        if target_display and title:
-            if target_display in title or title in target_display:
-                score += 4
-            elif target_display_tokens:
-                token_hits = sum(1 for tok in target_display_tokens if tok in title)
-                score += min(3, token_hits)
-        if score <= 0:
-            continue
-        candidates.append((score * 1000 - idx, tid))
-    if not candidates:
-        return False
-    candidates.sort(reverse=True)
-    selected_tid = int(candidates[0][1])
-
-    try:
-        player.mpv_set("sid", selected_tid)
-        try:
-            player.mpv_set("sub-visibility", True)
-        except Exception:
-            pass
-    except Exception:
-        return False
-
-    try:
-        updated = player.mpv_get("track-list")
-    except Exception:
-        updated = None
-    if isinstance(updated, list):
-        for row in updated:
-            if not isinstance(row, dict):
-                continue
-            if not _jellyfin_track_type_is_subtitle(row.get("type")):
-                continue
-            if not bool(row.get("selected")):
-                continue
-            src_id = None
-            ff_index = None
-            try:
-                src_id = int(row.get("src-id"))
-            except Exception:
-                src_id = None
-            try:
-                ff_index = int(row.get("ff-index"))
-            except Exception:
-                ff_index = None
-            if preferred_stream_index is not None:
-                if ff_index is not None and ff_index == preferred_stream_index:
-                    return True
-                if src_id is not None and (src_id - 1) == preferred_stream_index:
-                    return True
-            lang = _normalize_lang_pref(str(row.get("lang") or row.get("language") or ""))
-            title = str(row.get("title") or row.get("name") or "").strip().lower()
-            if target_lang and _language_matches(target_lang, lang):
-                return True
-            if target_display and target_display in title:
-                return True
-    return False
-
-
-def _jellyfin_runtime_selected_audio_stream(audio_streams: list[dict[str, object]]) -> tuple[int | None, str]:
-    """Resolve selected Jellyfin audio stream index from mpv runtime track data."""
-    try:
-        track_list = player.mpv_get("track-list")
-    except Exception:
-        track_list = None
-    if not isinstance(track_list, list):
-        return None, ""
-
-    selected: dict[str, object] | None = None
-    for row in track_list:
-        if not isinstance(row, dict):
-            continue
-        if str(row.get("type") or "").strip().lower() != "audio":
-            continue
-        if bool(row.get("selected")):
-            selected = row
-            break
-    if not isinstance(selected, dict):
-        return None, ""
-
-    selected_lang = _normalize_lang_pref(str(selected.get("lang") or selected.get("language") or ""))
-    selected_title = str(selected.get("title") or selected.get("name") or "").strip().lower()
-    selected_title_tokens = [tok for tok in re.split(r"[^a-z0-9]+", selected_title) if len(tok) >= 3]
-
-    try:
-        ff_index = int(selected.get("ff-index"))
-    except Exception:
-        ff_index = None
-    try:
-        src_id = int(selected.get("src-id"))
-    except Exception:
-        src_id = None
-
-    if ff_index is not None:
-        for row in audio_streams:
-            if not isinstance(row, dict):
-                continue
-            try:
-                if int(row.get("index")) == ff_index:
-                    return ff_index, selected_lang
-            except Exception:
-                continue
-    if src_id is not None:
-        candidate = int(src_id) - 1
-        for row in audio_streams:
-            if not isinstance(row, dict):
-                continue
-            try:
-                if int(row.get("index")) == candidate:
-                    return candidate, selected_lang
-            except Exception:
-                continue
-
-    best_idx: int | None = None
-    best_score = 0
-    for row in audio_streams:
-        if not isinstance(row, dict):
-            continue
-        try:
-            idx = int(row.get("index"))
-        except Exception:
-            continue
-        score = 0
-        row_lang = _normalize_lang_pref(str(row.get("language") or ""))
-        if selected_lang and _language_matches(selected_lang, row_lang):
-            score += 3
-        row_display = str(row.get("display") or "").strip().lower()
-        if selected_title and row_display:
-            if selected_title in row_display or row_display in selected_title:
-                score += 2
-            elif selected_title_tokens:
-                token_hits = sum(1 for tok in selected_title_tokens if tok in row_display)
-                score += min(2, token_hits)
-        if score > best_score:
-            best_score = score
-            best_idx = idx
-    return best_idx, selected_lang
-
-
-def _jellyfin_runtime_selected_subtitle_stream(subtitle_streams: list[dict[str, object]]) -> tuple[int | None, str, bool]:
-    """Resolve selected Jellyfin subtitle stream index from mpv runtime track data."""
-    try:
-        props = player.mpv_get_many(["track-list", "sid", "sub-visibility"])
-    except Exception:
-        props = {}
-    track_list = props.get("track-list") if isinstance(props, dict) else None
-    sid_raw = props.get("sid") if isinstance(props, dict) else None
-    sub_visible = props.get("sub-visibility") if isinstance(props, dict) else None
-    sid_text = str(sid_raw or "").strip().lower()
-    sid_off = sid_text in {"no", "0", "-1", "false"}
-    if sub_visible is False or sid_off:
-        return None, "off", True
-    if not isinstance(track_list, list):
-        return None, "", False
-
-    selected: dict[str, object] | None = None
-    for row in track_list:
-        if not isinstance(row, dict):
-            continue
-        if not _jellyfin_track_type_is_subtitle(row.get("type")):
-            continue
-        if bool(row.get("selected")):
-            selected = row
-            break
-    if not isinstance(selected, dict):
-        return None, "", False
-
-    selected_lang = _normalize_lang_pref(str(selected.get("lang") or selected.get("language") or ""))
-    selected_title = str(selected.get("title") or selected.get("name") or "").strip().lower()
-    selected_title_tokens = [tok for tok in re.split(r"[^a-z0-9]+", selected_title) if len(tok) >= 3]
-
-    try:
-        ff_index = int(selected.get("ff-index"))
-    except Exception:
-        ff_index = None
-    try:
-        src_id = int(selected.get("src-id"))
-    except Exception:
-        src_id = None
-
-    if ff_index is not None:
-        for row in subtitle_streams:
-            if not isinstance(row, dict):
-                continue
-            try:
-                if int(row.get("index")) == ff_index:
-                    return ff_index, selected_lang, False
-            except Exception:
-                continue
-    if src_id is not None:
-        candidate = int(src_id) - 1
-        for row in subtitle_streams:
-            if not isinstance(row, dict):
-                continue
-            try:
-                if int(row.get("index")) == candidate:
-                    return candidate, selected_lang, False
-            except Exception:
-                continue
-
-    best_idx: int | None = None
-    best_score = 0
-    for row in subtitle_streams:
-        if not isinstance(row, dict):
-            continue
-        try:
-            idx = int(row.get("index"))
-        except Exception:
-            continue
-        score = 0
-        row_lang = _normalize_lang_pref(str(row.get("language") or ""))
-        if selected_lang and _language_matches(selected_lang, row_lang):
-            score += 3
-        row_display = str(row.get("display") or "").strip().lower()
-        if selected_title and row_display:
-            if selected_title in row_display or row_display in selected_title:
-                score += 2
-            elif selected_title_tokens:
-                token_hits = sum(1 for tok in selected_title_tokens if tok in row_display)
-                score += min(2, token_hits)
-        if score > best_score:
-            best_score = score
-            best_idx = idx
-    return best_idx, selected_lang, False
-    return None, selected_lang
 
 
 def _jellyfin_emit_progress_hint() -> None:
