+++ tests/test_jellyfin_inventory.py
@@ -82,8 +82,9 @@ def test_jellyfin_inventory_doc_matches_source() -> None:
 # end-state residents; every leading-underscore product helper listed here is
 # migration inventory. Phase-start baseline: 117 definitions.
 EXPECTED_JELLYFIN_ROUTE_FUNCTIONS: dict[str, set[str]] = {
-    # M2 moved the 32 pure parse/normalize/URL helpers and M3 the stream
-    # selection/transcode policy and playable-item resolution to
+    # M2 moved the 32 pure parse/normalize/URL helpers, M3 the stream
+    # selection/transcode policy and playable-item resolution, and M4 the
+    # track preference/runtime track switching family to
     # integrations/jellyfin_service.py; routes keeps assignment aliases.
     "routes/__init__.py": {
         "_jellyfin_complete_ratio",
@@ -91,24 +92,16 @@ EXPECTED_JELLYFIN_ROUTE_FUNCTIONS: dict[str, set[str]] = {
         "_jellyfin_emit_progress_hint",
         "_jellyfin_emit_stopped_hint",
         "_jellyfin_emit_stopped_payload",
-        "_jellyfin_enrich_now_stream_metadata",
         "_jellyfin_integration_command_impl",
         "_jellyfin_is_duplicate_command",
         "_jellyfin_played_percentage",
         "_jellyfin_progress_snapshot",
-        "_jellyfin_runtime_selected_audio_stream",
-        "_jellyfin_runtime_selected_subtitle_stream",
         "_jellyfin_should_suppress_duplicate_play",
         "_jellyfin_should_suppress_duplicate_ui_action",
         "_jellyfin_snap_position_ticks",
         "_jellyfin_stopped_snapshot",
         "_jellyfin_stopped_snapshot_from_now",
-        "_jellyfin_try_set_mpv_audio_track",
-        "_jellyfin_try_set_mpv_subtitle_track",
-        "_merge_jellyfin_playback_metadata",
-        "_preferred_jellyfin_stream_indices",
         "_reset_jellyfin_command_state",
-        "_retarget_jellyfin_queue_stream_preferences",
         "_ui_event_push_jellyfin",
     },
     # Static asset endpoints — permanent HTTP surface, not migration targets.
