+++ tests/test_transition_inventory.py
@@ -127,6 +127,8 @@ def test_transition_inventory_doc_matches_source() -> None:
 #   (self-healing on read); candidate for a service reconcile command later.
 # - routes/queue.py, routes/uploads.py, upload_store.py: queue CRUD and
 #   upload retention — queue content management, not playback transitions.
+# - integrations/jellyfin_service.py: queue URL retargeting after language
+#   preference changes (Phase 4 M4) — content rewrite, not a transition.
 # - _TEMP_PLAYBACK_STACK compat aliases/wrappers: retained so existing tests
 #   can observe the live stack via the routes module.
 EXPECTED_TRANSITION_WRITERS: dict[str, set[str]] = {
@@ -145,6 +147,7 @@ EXPECTED_TRANSITION_WRITERS: dict[str, set[str]] = {
     "SESSION_POSITION": {"playback_service.py", "player.py"},
     "AUTO_NEXT_SUPPRESS_UNTIL": {"playback_service.py"},
     "QUEUE": {
+        "integrations/jellyfin_service.py",
         "playback_service.py",
         "player.py",
         "routes/__init__.py",
